fix(frontend): keep organization tab helpers out of backend typecheck
The unit test was importing the icon-backed tab constants, which made backend tsgo resolve ~icons modules.
