+++ src/layouts/settings.vue
@@ -10,8 +10,14 @@ import FailedCard from '~/components/FailedCard.vue'
 import RbacPermissionOnlyModal from '~/components/RbacPermissionOnlyModal.vue'
 import Tabs from '~/components/Tabs.vue'
 import { accountTabs } from '~/constants/accountTabs'
+import { organizationTabs as baseOrgTabs } from '~/constants/organizationTabs'
+import { settingsTabs } from '~/constants/settingsTabs'
+import { isNativeAppStoreContext } from '~/services/nativeCompliance'
+import { checkPermissions } from '~/services/permissions'
+import { openPortal } from '~/services/stripe'
+import { stripeEnabled } from '~/services/supabase'
+import { useOrganizationStore } from '~/stores/organization'
 import {
-  organizationTabs as baseOrgTabs,
   BILLING_TAB_KEY,
   cloneTabs,
   defaultChild,
@@ -19,13 +25,7 @@ import {
   findActiveTabKey,
   pathMatchesTab,
   TEAM_TAB_KEY,
-} from '~/constants/organizationTabs'
-import { settingsTabs } from '~/constants/settingsTabs'
-import { isNativeAppStoreContext } from '~/services/nativeCompliance'
-import { checkPermissions } from '~/services/permissions'
-import { openPortal } from '~/services/stripe'
-import { stripeEnabled } from '~/services/supabase'
-import { useOrganizationStore } from '~/stores/organization'
+} from '~/utils/organizationTabs'
 
 const { t } = useI18n()
 const organizationStore = useOrganizationStore()
